VZ-9459.  Reorganize validators in CertManager components - create a common package for shared helpers and test utils
