+++ platform-operator/controllers/verrazzano/component/certmanager/common/acme_utils.go
@@ -1,0 +1,19 @@
+// Copyright (c) 2023, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package common
+
+import (
+	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	"strings"
+)
+
+func IsLetsEncryptProductionEnv(acme v1beta1.Acme) bool {
+	return strings.ToLower(acme.Environment) == letsencryptProduction
+}
+
+func IsLetsEncryptStaging(compContext spi.ComponentContext) bool {
+	acmeEnvironment := compContext.EffectiveCR().Spec.Components.CertManager.Certificate.Acme.Environment
+	return acmeEnvironment != "" && strings.ToLower(acmeEnvironment) != "production"
+}
