+++ platform-operator/controllers/verrazzano/component/certmanager/common/constants.go
@@ -1,0 +1,19 @@
+// Copyright (c) 2023, Oracle and/or its affiliates.
+// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
+package common
+
+// CertManagerConfigComponentName is the name of the CertManager config component
+const CertManagerConfigComponentName = "cert-manager-config"
+
+// CertManagerComponentName is the name of the Verrazzano CertManager component
+const CertManagerComponentName = "cert-manager"
+
+// CertManagerComponentJSONName is the JSON name of the verrazzano component in CRD
+const CertManagerComponentJSONName = "certManager"
+
+// ExternalDNSComponentJSONName is the JSON name of the verrazzano component in CRD
+const ExternalDNSComponentJSONName = "externalCertManager"
+
+// CertManagerOCIDNSComponentName is the name of the OCI DNS webhook component
+const CertManagerOCIDNSComponentName = "verrazzano-ocidns-webhook"
