+++ platform-operator/controllers/verrazzano/component/certmanager/config/certmanager_config_component.go
@@ -6,6 +6,9 @@ package config
 import (
 	vzconst "github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
+	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
+	cmcommon "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/controller"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/ocidns"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
@@ -16,7 +19,7 @@ import (
 )
 
 // ComponentName is the name of the component
-const ComponentName = "cert-manager-config"
+const ComponentName = cmcommon.CertManagerConfigComponentName
 
 // ComponentNamespace is the namespace of the component
 const ComponentNamespace = vzconst.CertManagerNamespace
@@ -125,4 +128,62 @@ func (c certManagerConfigComponent) Uninstall(compContext spi.ComponentContext) 
 		return nil
 	}
 	return c.uninstallVerrazzanoCertManagerResources(compContext)
+}
+
+// ValidateInstall checks if the specified new Verrazzano CR is valid for this component to be installed
+func (c certManagerConfigComponent) ValidateInstall(vz *v1alpha1.Verrazzano) error {
+	vzV1Beta1 := &v1beta1.Verrazzano{}
+	if err := vz.ConvertTo(vzV1Beta1); err != nil {
+		return err
+	}
+	return c.ValidateInstallV1Beta1(vzV1Beta1)
+}
+
+// ValidateInstallV1Beta1 checks if the specified new Verrazzano CR is valid for this component to be installed
+func (c certManagerConfigComponent) ValidateInstallV1Beta1(vz *v1beta1.Verrazzano) error {
+	if !c.IsEnabled(vz) {
+		return nil
+	}
+
+	if err := cmcommon.ValidateLongestHostName(vz); err != nil {
+		return err
+	}
+
+	// Do not allow any changes except to enable the component post-install
+	cm := vz.Spec.Components.CertManager
+	if cm != nil {
+		if err := cmcommon.ValidateConfiguration(cm.Certificate); err != nil {
+			return err
+		}
+	}
+
+	return c.HelmComponent.ValidateInstallV1Beta1(vz)
+}
+
+// ValidateUpdate checks if the specified new Verrazzano CR is valid for this component to be updated
+func (c certManagerConfigComponent) ValidateUpdate(old *v1alpha1.Verrazzano, new *v1alpha1.Verrazzano) error {
+	oldBeta := &v1beta1.Verrazzano{}
+	newBeta := &v1beta1.Verrazzano{}
+	if err := old.ConvertTo(oldBeta); err != nil {
+		return err
+	}
+	if err := new.ConvertTo(newBeta); err != nil {
+		return err
+	}
+
+	return c.ValidateUpdateV1Beta1(oldBeta, newBeta)
+}
+
+// ValidateUpdateV1Beta1 checks if the specified new Verrazzano CR is valid for this component to be updated
+func (c certManagerConfigComponent) ValidateUpdateV1Beta1(old *v1beta1.Verrazzano, new *v1beta1.Verrazzano) error {
+	if !c.IsEnabled(new) {
+		return nil
+	}
+
+	cm := new.Spec.Components.CertManager
+	if err := cmcommon.ValidateConfiguration(cm.Certificate); err != nil {
+		return err
+	}
+
+	return c.HelmComponent.ValidateUpdateV1Beta1(old, new)
 }
