+++ platform-operator/controllers/verrazzano/component/certmanager/controller/certmanager_component_test.go
@@ -10,6 +10,7 @@ import (
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
+	cmcommon "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	corev1 "k8s.io/api/core/v1"
@@ -65,7 +66,6 @@ type validationTestStruct struct {
 
 var disabled = false
 
-const emailAddress = "joeblow@foo.com"
 const secretName = "newsecret"
 const secretNamespace = "ns"
 
@@ -122,133 +122,6 @@ var tests = []validationTestStruct{
 		wantErr: false,
 	},
 	{
-		name: "updateCustomCA",
-		old:  &vzapi.Verrazzano{},
-		new:  getCaSecretCR(),
-		caSecret: &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{Name: secretName, Namespace: secretNamespace},
-		},
-		wantErr: false,
-	},
-	{
-		name:    "updateCustomCASecretNotFound",
-		old:     &vzapi.Verrazzano{},
-		new:     getCaSecretCR(),
-		wantErr: true,
-	},
-	{
-		name:    "no change",
-		old:     &vzapi.Verrazzano{},
-		new:     &vzapi.Verrazzano{},
-		wantErr: false,
-	},
-	{
-		name: "updateInvalidBothConfigured",
-		old:  &vzapi.Verrazzano{},
-		new: &vzapi.Verrazzano{
-			Spec: vzapi.VerrazzanoSpec{
-				Components: vzapi.ComponentSpec{
-					CertManager: &vzapi.CertManagerComponent{
-						Certificate: vzapi.Certificate{
-							CA: vzapi.CA{
-								SecretName:               secretName,
-								ClusterResourceNamespace: secretNamespace,
-							},
-							Acme: vzapi.Acme{
-								Provider:     vzapi.LetsEncrypt,
-								EmailAddress: emailAddress,
-								Environment:  "staging",
-							},
-						},
-					},
-				},
-			},
-		},
-		wantErr: true,
-	},
-	{
-		name:    "validLetsEncryptStaging",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, "staging"),
-		wantErr: false,
-	},
-	{
-		name: "validLetsEncryptProviderCaseInsensitivity",
-		old:  &vzapi.Verrazzano{},
-		new: &vzapi.Verrazzano{
-			Spec: vzapi.VerrazzanoSpec{
-				Components: vzapi.ComponentSpec{
-					CertManager: &vzapi.CertManagerComponent{
-						Certificate: vzapi.Certificate{
-							Acme: vzapi.Acme{
-								Provider:     "LETSENCRYPT",
-								EmailAddress: emailAddress,
-								Environment:  letsEncryptStaging,
-							},
-						},
-					},
-				},
-			},
-		},
-		wantErr: false,
-	},
-	{
-		name:    "validLetsEncryptStagingCaseInsensitivity",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, "STAGING"),
-		wantErr: false,
-	},
-	{
-		name:    "validLetsEncryptProdCaseInsensitivity",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, "PRODUCTION"),
-		wantErr: false,
-	},
-	{
-		name:    "validLetsEncryptDefaultStagingEnv",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, ""),
-		wantErr: false,
-	},
-	{
-		name:    "validLetsEncryptProd",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, letsencryptProduction),
-		wantErr: false,
-	},
-	{
-		name: "invalidACMEProvider",
-		old:  &vzapi.Verrazzano{},
-		new: &vzapi.Verrazzano{
-			Spec: vzapi.VerrazzanoSpec{
-				Components: vzapi.ComponentSpec{
-					CertManager: &vzapi.CertManagerComponent{
-						Certificate: vzapi.Certificate{
-							Acme: vzapi.Acme{
-								Provider:     "blah",
-								EmailAddress: emailAddress,
-								Environment:  letsencryptProduction,
-							},
-						},
-					},
-				},
-			},
-		},
-		wantErr: true,
-	},
-	{
-		name:    "invalidLetsEncryptEnv",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, emailAddress, "myenv"),
-		wantErr: true,
-	},
-	{
-		name:    "invalidACMEEmail",
-		old:     &vzapi.Verrazzano{},
-		new:     getAcmeCR(vzapi.LetsEncrypt, "joeblow", letsEncryptStaging),
-		wantErr: true,
-	},
-	{
 		name: "singleOverride",
 		new:  getSingleOverrideCR(),
 		caSecret: &corev1.Secret{
@@ -273,14 +146,14 @@ func validationTests(t *testing.T, isUpdate bool) {
 				tt.wantErr = false
 			}
 			c := NewComponent()
-			getClientFunc = getTestClient(tt)
+			cmcommon.GetClientFunc = getTestClient(tt)
 			runValidationTest(t, tt, isUpdate, c)
 		})
 	}
 }
 
 func runValidationTest(t *testing.T, tt validationTestStruct, isUpdate bool, c spi.Component) {
-	defer func() { k8sutil.GetCoreV1Func = k8sutil.GetCoreV1Client }()
+	defer func() { cmcommon.ResetCoreV1ClientFunc() }()
 	if isUpdate {
 		if err := c.ValidateUpdate(tt.old, tt.new); (err != nil) != tt.wantErr {
 			t.Errorf("ValidateUpdate() error = %v, wantErr %v", err, tt.wantErr)
@@ -320,41 +193,6 @@ func getTestClient(tt validationTestStruct) func(log ...vzlog.VerrazzanoLogger) 
 			return createFakeClient(tt.caSecret).CoreV1(), nil
 		}
 		return createFakeClient().CoreV1(), nil
-	}
-}
-
-func getAcmeCR(provider vzapi.ProviderType, emailAddr string, env string) *vzapi.Verrazzano {
-	return &vzapi.Verrazzano{
-		Spec: vzapi.VerrazzanoSpec{
-			Components: vzapi.ComponentSpec{
-				CertManager: &vzapi.CertManagerComponent{
-					Certificate: vzapi.Certificate{
-						Acme: vzapi.Acme{
-							Provider:     provider,
-							EmailAddress: emailAddr,
-							Environment:  env,
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func getCaSecretCR() *vzapi.Verrazzano {
-	return &vzapi.Verrazzano{
-		Spec: vzapi.VerrazzanoSpec{
-			Components: vzapi.ComponentSpec{
-				CertManager: &vzapi.CertManagerComponent{
-					Certificate: vzapi.Certificate{
-						CA: vzapi.CA{
-							SecretName:               secretName,
-							ClusterResourceNamespace: secretNamespace,
-						},
-					},
-				},
-			},
-		},
 	}
 }
