+++ platform-operator/controllers/verrazzano/component/certmanager/controller/certmanager_test.go
@@ -5,25 +5,14 @@ package controller
 
 import (
 	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/pem"
-	"fmt"
-	"math/big"
-	"net"
-	"testing"
-	"time"
-
 	certv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/verrazzano/verrazzano/pkg/bom"
 	"github.com/verrazzano/verrazzano/pkg/constants"
-	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/common"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
 	appsv1 "k8s.io/api/apps/v1"
@@ -37,13 +26,13 @@ import (
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	clipkg "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+	"testing"
 )
 
 const (
-	profileDir      = "../../../../../manifests/profiles"
-	testNamespace   = "testNamespace"
-	testBomFile     = "../../../testdata/test_bom.json"
-	fooDomainSuffix = "foo.com"
+	profileDir    = "../../../../../manifests/profiles"
+	testNamespace = "testNamespace"
+	testBomFile   = "../../../testdata/test_bom.json"
 )
 
 const (
@@ -56,69 +45,10 @@ const (
 	compartmentID    = "compartmentID"
 )
 
-// Default Verrazzano object
-var vz = &vzapi.Verrazzano{
-	ObjectMeta: metav1.ObjectMeta{Name: myvz, Namespace: myvzns, CreationTimestamp: metav1.Now()},
-	Spec: vzapi.VerrazzanoSpec{
-		EnvironmentName: "myenv",
-		Components: vzapi.ComponentSpec{
-			DNS: &vzapi.DNSComponent{},
-		},
-	},
-}
-
-// Default Verrazzano v1beta1 object
-var vzv1beta1 = &v1beta1.Verrazzano{
-	ObjectMeta: metav1.ObjectMeta{Name: myvz, Namespace: myvzns, CreationTimestamp: metav1.Now()},
-	Spec: v1beta1.VerrazzanoSpec{
-		EnvironmentName: "myenv",
-		Components: v1beta1.ComponentSpec{
-			DNS: &v1beta1.DNSComponent{},
-		},
-	},
-}
-
-var oci = &vzapi.OCI{
-	OCIConfigSecret:        ociDNSSecretName,
-	DNSZoneCompartmentOCID: compartmentID,
-	DNSZoneOCID:            zoneID,
-	DNSZoneName:            zoneName,
-}
-
-var ociV1Beta1 = &v1beta1.OCI{
-	OCIConfigSecret:        ociDNSSecretName,
-	DNSZoneCompartmentOCID: compartmentID,
-	DNSZoneOCID:            zoneID,
-	DNSZoneName:            zoneName,
-}
-
-var ociLongDNSZoneName = &vzapi.OCI{
-	OCIConfigSecret:        ociDNSSecretName,
-	DNSZoneCompartmentOCID: compartmentID,
-	DNSZoneOCID:            zoneID,
-	DNSZoneName:            "veryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryverylong.name.io",
-	DNSScope:               "#jhwuyusj!!!",
-}
-
-var ociLongDNSZoneNameV1Beta1 = &v1beta1.OCI{
-	OCIConfigSecret:        ociDNSSecretName,
-	DNSZoneCompartmentOCID: compartmentID,
-	DNSZoneOCID:            zoneID,
-	DNSZoneName:            "veryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryverylong.name.io",
-	DNSScope:               "#jhwuyusj!!!",
-}
-
 // default CA object
 var ca = vzapi.CA{
 	SecretName:               "testSecret",
 	ClusterResourceNamespace: testNamespace,
-}
-
-// Default Acme object
-var acme = vzapi.Acme{
-	Provider:     vzapi.LetsEncrypt,
-	EmailAddress: "testEmail@foo.com",
-	Environment:  letsEncryptStaging,
 }
 
 // Default Verrazzano object
@@ -150,45 +80,8 @@ func init() {
 // THEN the function returns true
 func TestIsCertManagerEnabled(t *testing.T) {
 	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Enabled = getBoolPtr(true)
+	localvz.Spec.Components.CertManager.Enabled = common.GetBoolPtr(true)
 	assert.True(t, fakeComponent.IsEnabled(spi.NewFakeContext(nil, localvz, nil, false).EffectiveCR()))
-}
-
-// TestIsOCIDNS tests whether the Effective CR is using OCI DNS
-// GIVEN a call to isOCIDNS
-// WHEN OCI DNS is specified in the Verrazzano Spec
-// THEN isOCIDNS should return true
-func TestIsOCIDNS(t *testing.T) {
-	var tests = []struct {
-		name   string
-		vz     *vzapi.Verrazzano
-		ocidns bool
-	}{
-		{
-			"shouldn't be enabled when nil",
-			&vzapi.Verrazzano{},
-			false,
-		},
-		{
-			"should be enabled when OCI DNS present",
-			&vzapi.Verrazzano{
-				Spec: vzapi.VerrazzanoSpec{
-					Components: vzapi.ComponentSpec{
-						DNS: &vzapi.DNSComponent{
-							OCI: &vzapi.OCI{},
-						},
-					},
-				},
-			},
-			true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.ocidns, isOCIDNS(tt.vz))
-		})
-	}
 }
 
 // TestWriteCRD tests writing out the OCI DNS metadata to CertManager CRDs
@@ -216,7 +109,7 @@ func TestCleanTempFiles(t *testing.T) {
 // THEN the function returns false
 func TestIsCertManagerDisabled(t *testing.T) {
 	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Enabled = getBoolPtr(false)
+	localvz.Spec.Components.CertManager.Enabled = common.GetBoolPtr(false)
 	assert.False(t, fakeComponent.IsEnabled(spi.NewFakeContext(nil, localvz, nil, false).EffectiveCR()))
 }
 
@@ -238,8 +131,9 @@ func TestAppendCertManagerOverrides(t *testing.T) {
 func TestAppendCertManagerOverridesWithInstallArgs(t *testing.T) {
 	localvz := defaultVZConfig.DeepCopy()
 	localvz.Spec.Components.CertManager.Certificate.CA = ca
-	defer func() { getClientFunc = k8sutil.GetCoreV1Client }()
-	getClientFunc = createClientFunc(localvz.Spec.Components.CertManager.Certificate.CA, "defaultVZConfig-cn")
+	defer func() { common.ResetCoreV1ClientFunc() }()
+	common.GetClientFunc = createClientFunc(localvz.Spec.Components.CertManager.Certificate.CA, "defaultVZConfig-cn")
+
 	kvs, err := AppendOverrides(spi.NewFakeContext(nil, localvz, nil, false), ComponentName, ComponentNamespace, "", []bom.KeyValue{})
 	assert.NoError(t, err)
 	assert.Len(t, kvs, 1)
@@ -344,36 +238,7 @@ func TestIsCertManagerNotReady(t *testing.T) {
 	assert.False(t, certManager.isCertManagerReady(spi.NewFakeContext(client, nil, nil, false)))
 }
 
-// TestIsCANilWithProfile tests the isCA function
-// GIVEN a call to isCA
-// WHEN the CertManager component is populated by the profile
-// THEN true is returned
-func TestIsCANilWithProfile(t *testing.T) {
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-	isCAValue, err := IsCA(spi.NewFakeContext(client, &vzapi.Verrazzano{}, nil, false, profileDir))
-	assert.Nil(t, err)
-	assert.True(t, isCAValue)
-}
-
-// TestIsCANilTrue tests the isCA function
-// GIVEN a call to isCA
-// WHEN the Certificate CA is populated
-// THEN true is returned
-func TestIsCATrue(t *testing.T) {
-	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.CA = ca
-
-	defer func() { getClientFunc = k8sutil.GetCoreV1Client }()
-	getClientFunc = createClientFunc(localvz.Spec.Components.CertManager.Certificate.CA, "defaultVZConfig-cn")
-
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-
-	isCAValue, err := IsCA(spi.NewFakeContext(client, localvz, nil, false, profileDir))
-	assert.Nil(t, err)
-	assert.True(t, isCAValue)
-}
-
-func createClientFunc(caConfig vzapi.CA, cn string, otherObjs ...runtime.Object) getCoreV1ClientFuncType {
+func createClientFunc(caConfig vzapi.CA, cn string, otherObjs ...runtime.Object) common.GetCoreV1ClientFuncType {
 	return func(...vzlog.VerrazzanoLogger) (corev1.CoreV1Interface, error) {
 		secret, err := createCertSecretNoParent(caConfig.SecretName, caConfig.ClusterResourceNamespace, cn)
 		if err != nil {
@@ -383,32 +248,6 @@ func createClientFunc(caConfig vzapi.CA, cn string, otherObjs ...runtime.Object)
 		objs = append(objs, otherObjs...)
 		return k8sfake.NewSimpleClientset(objs...).CoreV1(), nil
 	}
-}
-
-// TestIsCANilFalse tests the isCA function
-// GIVEN a call to isCA
-// WHEN the Certificate Acme is populated
-// THEN false is returned
-func TestIsCAFalse(t *testing.T) {
-	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.Acme = acme
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-	isCAValue, err := IsCA(spi.NewFakeContext(client, localvz, nil, false, profileDir))
-	assert.Nil(t, err)
-	assert.False(t, isCAValue)
-}
-
-// TestIsCANilFalse tests the isCA function
-// GIVEN a call to isCA
-// WHEN the Certificate Acme is populated
-// THEN false is returned
-func TestIsCABothPopulated(t *testing.T) {
-	localvz := defaultVZConfig.DeepCopy()
-	localvz.Spec.Components.CertManager.Certificate.CA = ca
-	localvz.Spec.Components.CertManager.Certificate.Acme = acme
-	client := fake.NewClientBuilder().WithScheme(testScheme).Build()
-	_, err := IsCA(spi.NewFakeContext(client, localvz, nil, false, profileDir))
-	assert.Error(t, err)
 }
 
 // TestUninstallCertManager tests the cert-manager uninstall process
@@ -549,120 +388,6 @@ func TestGetOverrides(t *testing.T) {
 	}
 }
 
-// TestValidateLongestHostName tests the following scenarios
-// GIVEN a call to validateLongestHostName func
-// WHEN the CR passed is v1alpha1
-// THEN it is inspected to validate the host name length of endpoints
-func TestValidateLongestHostName(t *testing.T) {
-	asserts := assert.New(t)
-	cr1, cr2, cr3, cr4, cr5, cr6 := *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy(), *vz.DeepCopy()
-	cr1.Spec.Components.DNS.OCI = ociLongDNSZoneName
-	cr2.Spec.Components.DNS.OCI = oci
-	cr3.Spec.EnvironmentName = "veryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryverylong"
-	cr3.Spec.Components.DNS = nil
-	cr4.Spec.Components.DNS = nil
-	cr5.Spec.Components.DNS = &vzapi.DNSComponent{External: &vzapi.External{Suffix: ociLongDNSZoneNameV1Beta1.DNSZoneName}}
-	cr6.Spec.Components.DNS = &vzapi.DNSComponent{External: &vzapi.External{Suffix: fooDomainSuffix}}
-	tests := []struct {
-		cr        vzapi.Verrazzano
-		wantError bool
-		want      string
-	}{
-		{
-			cr:        cr1,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s and DNS suffix %s are too long. For the given configuration they must have at most %v characters in combination", cr1.Spec.EnvironmentName, cr1.Spec.Components.DNS.OCI.DNSZoneName, 64-preOccupiedspace),
-		},
-		{
-			cr:        cr2,
-			wantError: false,
-		},
-		{
-			cr:        cr3,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s is too long. For the given configuration it must have at most %v characters", cr3.Spec.EnvironmentName, 64-(14+preOccupiedspace)),
-		},
-		{
-			cr:        cr4,
-			wantError: false,
-		},
-		{
-			cr:        cr5,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s and DNS suffix %s are too long. For the given configuration they must have at most %v characters in combination", cr5.Spec.EnvironmentName, cr5.Spec.Components.DNS.External.Suffix, 64-preOccupiedspace),
-		},
-		{
-			cr:        cr6,
-			wantError: false,
-		},
-	}
-	for _, test := range tests {
-		err := validateLongestHostName(&test.cr)
-		if test.wantError {
-			asserts.EqualError(err, test.want)
-		} else {
-			asserts.NoError(err)
-		}
-	}
-}
-
-// TestValidateLongestHostNameV1Beta1 tests the following scenarios
-// GIVEN a call to validateLongestHostName func
-// WHEN the CR passed is v1beta1
-// THEN it is inspected to validate the host name length of endpoints
-func TestValidateLongestHostNameV1Beta1(t *testing.T) {
-	asserts := assert.New(t)
-	cr1, cr2, cr3, cr4, cr5, cr6 := *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy(), *vzv1beta1.DeepCopy()
-	cr1.Spec.Components.DNS.OCI = ociLongDNSZoneNameV1Beta1
-	cr2.Spec.Components.DNS.OCI = ociV1Beta1
-	cr3.Spec.EnvironmentName = "veryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryveryverylong"
-	cr3.Spec.Components.DNS = nil
-	cr4.Spec.Components.DNS = nil
-	cr5.Spec.Components.DNS = &v1beta1.DNSComponent{External: &v1beta1.External{Suffix: ociLongDNSZoneNameV1Beta1.DNSZoneName}}
-	cr6.Spec.Components.DNS = &v1beta1.DNSComponent{External: &v1beta1.External{Suffix: fooDomainSuffix}}
-	tests := []struct {
-		cr        v1beta1.Verrazzano
-		wantError bool
-		want      string
-	}{
-		{
-			cr:        cr1,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s and DNS suffix %s are too long. For the given configuration they must have at most %v characters in combination", cr1.Spec.EnvironmentName, cr1.Spec.Components.DNS.OCI.DNSZoneName, 64-preOccupiedspace),
-		},
-		{
-			cr:        cr2,
-			wantError: false,
-		},
-		{
-			cr:        cr3,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s is too long. For the given configuration it must have at most %v characters", cr3.Spec.EnvironmentName, 64-(14+preOccupiedspace)),
-		},
-		{
-			cr:        cr4,
-			wantError: false,
-		},
-		{
-			cr:        cr5,
-			wantError: true,
-			want:      fmt.Sprintf("spec.environmentName %s and DNS suffix %s are too long. For the given configuration they must have at most %v characters in combination", cr5.Spec.EnvironmentName, cr5.Spec.Components.DNS.External.Suffix, 64-preOccupiedspace),
-		},
-		{
-			cr:        cr6,
-			wantError: false,
-		},
-	}
-	for _, test := range tests {
-		err := validateLongestHostName(&test.cr)
-		if test.wantError {
-			asserts.EqualError(err, test.want)
-		} else {
-			asserts.NoError(err)
-		}
-	}
-}
-
 // Create a new deployment object for testing
 func newDeployment(name string, labels map[string]string, updated bool) *appsv1.Deployment {
 	deployment := &appsv1.Deployment{
@@ -720,7 +445,7 @@ func newReplicaSet(name string) *appsv1.ReplicaSet {
 }
 
 func createCertSecretNoParent(name string, namespace string, cn string) (*v1.Secret, error) {
-	fakeCertBytes, err := createFakeCertBytes(cn, nil)
+	fakeCertBytes, err := common.CreateFakeCertBytes(cn, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -736,72 +461,4 @@ func createCertSecretNoParent(name string, namespace string, cn string) (*v1.Sec
 		Type: v1.SecretTypeTLS,
 	}
 	return secret, nil
-}
-
-var testRSAKey *rsa.PrivateKey
-
-func getRSAKey() (*rsa.PrivateKey, error) {
-	if testRSAKey == nil {
-		var err error
-		if testRSAKey, err = rsa.GenerateKey(rand.Reader, 2048); err != nil {
-			return nil, err
-		}
-	}
-	return testRSAKey, nil
-}
-
-func createFakeCertBytes(cn string, parent *x509.Certificate) ([]byte, error) {
-	rsaKey, err := getRSAKey()
-	if err != nil {
-		return []byte{}, err
-	}
-
-	cert := createFakeCertificate(cn)
-	if parent == nil {
-		parent = cert
-	}
-	pubKey := &rsaKey.PublicKey
-	certBytes, err := x509.CreateCertificate(rand.Reader, cert, parent, pubKey, rsaKey)
-	if err != nil {
-		return []byte{}, err
-	}
-	certPem := pem.EncodeToMemory(&pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: certBytes,
-	})
-	return certPem, nil
-}
-
-func createFakeCertificate(cn string) *x509.Certificate {
-	duration30, _ := time.ParseDuration("-30h")
-	notBefore := time.Now().Add(duration30) // valid 30 hours ago
-	duration1Year, _ := time.ParseDuration("90h")
-	notAfter := notBefore.Add(duration1Year) // for 90 hours
-	serialNo := big.NewInt(int64(123123413123))
-	cert := &x509.Certificate{
-		Subject: pkix.Name{
-			Country:      []string{"US"},
-			Organization: []string{"BarOrg"},
-			SerialNumber: "2234",
-			CommonName:   cn,
-		},
-		SerialNumber:          serialNo,
-		NotBefore:             notBefore,
-		NotAfter:              notAfter,
-		KeyUsage:              x509.KeyUsageKeyEncipherment,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-		BasicConstraintsValid: true,
-		IsCA:                  false,
-		SignatureAlgorithm:    x509.SHA256WithRSA,
-	}
-
-	cert.IPAddresses = append(cert.IPAddresses, net.ParseIP("127.0.0.1"))
-	cert.IPAddresses = append(cert.IPAddresses, net.ParseIP("::"))
-	cert.DNSNames = append(cert.DNSNames, "localhost")
-	return cert
-}
-
-// Create a bool pointer
-func getBoolPtr(b bool) *bool {
-	return &b
 }
